pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "data-assembler"
version = "0.1.0"
description = "Automated reflectivity data ingestion workflow for scientific data lakehouse"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Data Lakehouse Team" }
]
keywords = ["reflectometry", "neutron", "data-ingestion", "lakehouse"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Physics",
]
dependencies = [
    "pydantic>=2.0",
    "pyarrow>=14.0",
    "pandas>=2.0",
    "periodictable>=1.6",
    "numpy>=1.24",
    "python-dateutil>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
]
ai = [
    "openai>=1.0",
    "anthropic>=0.20",
]
iceberg = [
    "pyiceberg>=0.5.0",
]

[project.scripts]
data-assembler = "assembler.cli.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/assembler"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311", "py312"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
